# splinterd/api/static/openapi.yaml
...
openapi: '3.0.0'
info:
  version: 0.5.1
  title: splinterd API
  description: REST API for the Splinter daemon
servers:
  - url: http://localhost:9000/api
paths:
...
  /registry/nodes/{identity}:
    get:
      summary: Fetch a node in the registry by its identity
      description: |
        This endpoint can be used to view a specific nodes in the registry.
      tags:
        - Splinter Registry
      parameters:
        - $ref: "#/components/parameters/auth"
        - $ref: "#/components/parameters/protocol_version"
        - name: identity
          in: path
          description: identity of node to fetch
          required: true
          schema:
            type: string
      responses:
        200:
          description: The node was successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/RegisteredNode"
        401:
          description: The client is unauthorized
        404:
          description: The node was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: An internal server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Add or replace a node in the registry
      description: |
        This endpoint can be used to add a new node to the registry, or replace
        an existing node. When replacing an existing node, the node identity
        cannot be changed. This action is idempotent.
      tags:
        - Splinter Registry
      parameters:
        - $ref: "#/components/parameters/auth"
        - $ref: "#/components/parameters/protocol_version"
        - name: identity
          in: path
          description: identity of node to add/replace
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisteredNode'
      responses:
        200:
          description: The node has been added or replaced
        400:
          description: The request was malformed or the node was invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: The client is unauthorized
        500:
          description: An internal server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a node from the registry
      description: This endpoint can be used to remove a node from the registry.
      tags:
        - Splinter Registry
      parameters:
        - $ref: "#/components/parameters/auth"
        - $ref: "#/components/parameters/protocol_version"
        - name: identity
          in: path
          description: identity of node to delete
          required: true
          schema:
            type: string
      responses:
        200:
          description: The node has been deleted from the registry
        401:
          description: The client is unauthorized
        404:
          description: The node was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: An internal server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    auth:
      name: Authorization
      in: header
      description: |
        The client's authorization, which the server resolves to an identity.
        Currently supports Biome JWT (if Biome credentials is enabled), Cylinder
        JWT, and OAuth2 bearer tokens.
      required: true
      schema:
        type: string
        examples:
          Biome: Bearer Biome:eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lk\
            IjoiZjM1YWFjYzEtYTljZC00ZWRhLWI2ZDAtMmVmYWRkZjBjOGE0IiwiaXNzIjoic2V\
            sZi1pc3N1ZWQiLCJleHAiOjE1ODAyMzkyMjh9.P8hA0ru_xriYX7qryl08ZEp86t5HD\
            _AEVPEUXY70Ehc
          Cylinder: Bearer Cylinder:eyJhbGciOiJzZWNwMjU2azEiLCJ0eXAiOiJjeWxpbmR\
            lcitqd3QifQ==.eyJpc3MiOiIwMjc5YmU2NjdlZjlkY2JiYWM1NWEwNjI5NWNlODcwY\
            jA3MDI5YmZjZGIyZGNlMjhkOTU5ZjI4MTViMTZmODE3OTgifQ==.71Vw3+m9R4b8iZU\
            zhRHOAtX/hO5WYA9PgMe27/CeZ6NhIXFYkBBzreoIGpHbfJ8UxT+1MLUgjsQB8TISaf\
            neRA==
          OAuth2: Bearer OAuth2:55eeea7ce2b472d69d406990939baa698e34e955
    protocol_version:
      name: SplinterProtocolVersion
      in: header
      description: |
        The protocol version which the client can understand. If not provided,
        the node will respond using its latest protocol version.
      required: false
      schema:
        type: integer
        example: 2
  schemas:
    Error:
      additionalProperties: false
      properties:
        message:
          description: A message describing the error that occurred
          type: string
          example: DatabaseError({description})
      required:
        - message
...
    RegisteredNode:
      type: object
      properties:
        identity:
          type: string
        endpoints:
          type: array
          items:
            type: string
        display_name:
          type: string
        keys:
          type: array
          items:
            type: string
        metadata:
          type: object
      example:
        identity: node-123123-asdf
        endpoints:
          - tcps://12.0.0.123:8431
        display_name: Cargill - Node 1
        keys:
          - "03e0e5086beffc640ec0d149d4f1197fdde0f338afac774541831281c6fd91cbe0"
        metadata:
          company: Cargill
          status: Up
...